requirements.txt
cohere
python-dotenv
pandas
langchain